completed majority of the initial refactor
Majority of the initial refactor has been completed, which includes the redesign of the file structure of the project, aswell as the inclusion additonal plugins.

- addtional plugins:
  - todo-comments
  - alpha
  - neoclip
  - bufferline

Still need to complete the creation of all desired key-bindings for the additional modules and update documentation respectively.
