+++ lua/.luarc.json
@@ -1,0 +1,3 @@
+{
+    "workspace.checkThirdParty": false
+}
