+++ lua/core/init.lua
@@ -1,0 +1,2 @@
+require("core.keybindings")
+require("core.vim")
