+++ lua/extensions/init.lua
@@ -1,0 +1,1 @@
+require("extensions.lazygit")
