+++ lua/plugins/clipboard/init.lua
@@ -1,0 +1,1 @@
+require("plugins.clipboard.neoclip")
