+++ lua/plugins/clipboard/neoclip.lua
@@ -1,0 +1,4 @@
+-- TODO: Move to neoclip setup
+require('neoclip').setup({})
+require('telescope').load_extension('neoclip')
+
