+++ lua/plugins/cmp/init.lua
@@ -1,0 +1,1 @@
+require("plugins.cmp.cmp")
