+++ lua/plugins/dap/neodev.lua
@@ -1,0 +1,5 @@
+--! dependency of dapui
+require("neodev").setup({
+  library = { plugins = { "nvim-dap-ui" }, types = true },
+})
+
