+++ lua/plugins/git/diffview.lua
@@ -1,0 +1,3 @@
+-- default config setup
+require("diffview").setup({})
+
