+++ lua/plugins/git/gitsigns.lua
@@ -1,0 +1,11 @@
+--* --------------------------------------------------------------- *--
+--?                Git Workflow Related Configurations              ?--
+--* --------------------------------------------------------------- *--
+
+-- override gitsigns to reflect new lines with a 'plus' sign
+require("gitsigns").setup({
+  signs = {
+    add = { text = "+" }
+  }
+})
+
