+++ lua/plugins/git/init.lua
@@ -1,0 +1,3 @@
+require("plugins.git.gitsigns")
+require("plugins.git.diffview")
+require("plugins.git.neogit")
