+++ lua/plugins/git/neogit.lua
@@ -1,0 +1,3 @@
+-- default config setup
+require("neogit").setup({})
+
