+++ lua/plugins/icons/init.lua
@@ -1,0 +1,2 @@
+require("plugins.icons.nvim-web-devicons")
+require("plugins.icons.codicons")
