+++ lua/plugins/init.lua
@@ -1,0 +1,16 @@
+require("plugins.bufferline")
+-- BUG: there is an issue with shade, where it will not identifty popup window's overlay positions correctly, such as with `jackmort/chatgpt.nvim` - disabled until it can be corrected
+--require("plugins.shade")
+require("plugins.lualine")
+require("plugins.toggleterm")
+require("plugins.nvim-treesitter")
+require("plugins.nvim-comment")
+require("plugins.todo-comments")
+require("plugins.git")
+require("plugins.chatgpt")
+require("plugins.icons")
+require("plugins.dap")
+require("plugins.clipboard")
+require("plugins.lsp")
+require("plugins.cmp")
+require("plugins.ui")
