+++ lua/plugins/lsp/init.lua
@@ -1,0 +1,2 @@
+require("plugins.lsp.lsp-zero")
+require("plugins.lsp.lspconfig")
