+++ lua/plugins/nvim-comment.lua
@@ -1,0 +1,7 @@
+--* ------------------------------------ *--
+--?              nvim-comment            ?--
+--* ------------------------------------ *--
+require("nvim_comment").setup({
+  operator_mapping = "<leader>/"
+})
+
