+++ lua/plugins/ui/colorscheme.lua
@@ -1,0 +1,10 @@
+--* ------------------------------------ *--
+--?          Color Scheme Setup          ?--
+--* ------------------------------------ *--
+
+-- set the vim background to Dark Mode
+vim.o.background = "dark"
+
+-- set colorscheme to gruvbox, AFTER setting all configurations
+vim.cmd("colorscheme gruvbox")
+
