+++ lua/plugins/ui/init.lua
@@ -1,0 +1,4 @@
+require("plugins.ui.gruvbox")
+-- NOTE:the colorscheme needs to be imported after gruvbox
+require("plugins.ui.colorscheme")
+require("plugins.ui.alpha")
