+++ lua/utils/init.lua
@@ -1,0 +1,7 @@
+-- import the Helpers submodule
+local Helpers = require("utils.helpers")
+
+-- export the sub module
+return {
+  Helpers = Helpers
+}
